requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
supabase==2.0.3
pdfplumber==0.10.2
python-docx==1.0.1
openai==1.3.5
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
PyJWT==2.8.0
requests==2.31.0
PyJWT==2.8.0
requests==2.31.0

